Switches from Alfred to Raycast (#11)
TL;DR
-----

Stops installing Alfred, starts installing Raycast

Details
-------

I decided to jump on the Raycast bandwagon since I realized it had
clipboard history as well as the Alfred-like things it does. Flycut
and Maccy weren't super reliable for me lately, so I decided to
make the switch.

This change removes the Alfred cask, adds the one for Raycast,
and also drops the one for Flycut.